DSD/requirements.txt
absl-py==0.6.1
asn1crypto==0.24.0
astor==0.7.1
bleach==1.5.0
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
cryptography==2.4.2
cycler==0.10.0
Cython==0.29.1
Django==1.11.11
djangorestframework==3.9.0
gast==0.2.0
grpcio==1.16.1
h5py==2.8.0
html5lib==0.9999999
idna==2.7
Keras==2.2.4
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
keras-resnet==0.1.0
keras-retinanet==0.5.0
kiwisolver==1.0.1
Markdown==3.0.1
matplotlib==3.0.2
numpy==1.15.4
opencv-python==3.4.4.19
Pillow==5.3.0
progressbar2==3.38.0
protobuf==3.6.1
pycparser==2.19
PyMySQL==0.9.2
pyparsing==2.3.0
python-dateutil==2.7.5
python-utils==2.3.0
pytz==2018.7
PyYAML==3.13
requests==2.20.1
scipy==1.1.0
six==1.11.0
tensorboard==1.12.0
tensorflow==1.12.0
tensorflow-gpu==1.3.0
tensorflow-tensorboard==0.1.8
termcolor==1.1.0
urllib3==1.24.1
Werkzeug==0.14.1
